refactor: rename to cave (singular)
