+++ src/modules/actors/tiles/cave/cave-tile.factory.ts
